refactor: input 로직 분리
